Depricated uint8

+++ contracts/variableLabsEscrow.sol
@@ -106,7 +106,7 @@ contract variableLabsEscrow{
 		address receiver;
 		address resolver;
 		uint256 fee;
-		uint8 state;
+		uint256 state;
 		bool exists;
 	}
 
@@ -151,12 +151,12 @@ contract variableLabsEscrow{
 		currentEscrow.receiver = _receiver;
 		currentEscrow.resolver = _resolver;
 		currentEscrow.fee = _fee;
-		currentEscrow.state = uint8(0);
+		currentEscrow.state = 0;
 		currentEscrow.exists = true;
 
 		// Sender must approve this contract to spend the said amount of funds before it can be transfered.
-		uint256 _feeAmount = escrow[_id].funds.mul(uint256(escrow[_id].fee));
-		_feeAmount = _feeAmount.div(uint256(10000));
+		uint256 _feeAmount = escrow[_id].funds.mul(escrow[_id].fee);
+		_feeAmount = _feeAmount.div(10000);
 		Token(tokenAddress).transferFrom(msg.sender, address(this), (_funds+_feeAmount));
 
 		escrow[_id] = currentEscrow;
@@ -173,15 +173,15 @@ contract variableLabsEscrow{
 		if(!escrow[_id].exists) 
 			revert("escrow does not exist");
 		require(escrow[_id].depositer == msg.sender, "only by depositer");
-		require(escrow[_id].state == uint8(0), "escrow must be active");
+		require(escrow[_id].state == 0, "escrow must be active");
 
 		Token(tokenAddress).transfer(escrow[_id].receiver, escrow[_id].funds);
 
-		uint256 _feeAmount = escrow[_id].funds.mul(uint256(escrow[_id].fee));
-		_feeAmount = _feeAmount.div(uint256(10000));
+		uint256 _feeAmount = escrow[_id].funds.mul(escrow[_id].fee);
+		_feeAmount = _feeAmount.div(10000);
 		Token(tokenAddress).transfer(escrow[_id].resolver, _feeAmount);
 
-		escrow[_id].state = uint8(1);
+		escrow[_id].state = 1;
 
 		return true;
 	}
@@ -195,16 +195,16 @@ contract variableLabsEscrow{
 		if(!escrow[_id].exists) 
 			revert("escrow does not exist");
 		require(escrow[_id].receiver == msg.sender, "only by receiver");
-		require(escrow[_id].state == uint8(0), "escrow must be active");
+		require(escrow[_id].state == 0, "escrow must be active");
 		
 
 		Token(tokenAddress).transfer(escrow[_id].depositer, escrow[_id].funds);
 
-		uint256 _feeAmount = escrow[_id].funds.mul(uint256(escrow[_id].fee));
-		_feeAmount = _feeAmount.div(uint256(10000));
+		uint256 _feeAmount = escrow[_id].funds.mul(escrow[_id].fee);
+		_feeAmount = _feeAmount.div(10000);
 		Token(tokenAddress).transfer(escrow[_id].resolver, _feeAmount);
 
-		escrow[_id].state = uint8(2);
+		escrow[_id].state = 2;
 
 		return true;
 	}
@@ -218,13 +218,13 @@ contract variableLabsEscrow{
 		if(!escrow[_id].exists) 
 			revert("escrow does not exist");
 		require((escrow[_id].depositer == msg.sender) || (escrow[_id].receiver == msg.sender), "only by depositer or receiver");
-		require(escrow[_id].state == uint8(0), "escrow must be active");
+		require(escrow[_id].state == 0, "escrow must be active");
 
 		if(msg.sender == escrow[_id].depositer){
-			escrow[_id].state = uint8(3);
+			escrow[_id].state = 3;
 		}
 		if(msg.sender == escrow[_id].receiver){
-			escrow[_id].state = uint8(4);
+			escrow[_id].state = 4;
 		}
 
 		return true;
@@ -235,28 +235,28 @@ contract variableLabsEscrow{
 	@param _id @param _decision
 	Resolver resolves a disputed contract in either direction. Funds and fees are transferred.
     */
-	function resolveDispute(bytes32 _id, uint8 _decision) public returns (bool success){
+	function resolveDispute(bytes32 _id, uint256 _decision) public returns (bool success){
 		if(!escrow[_id].exists) 
 			revert("escrow does not exist");
-		require((escrow[_id].state == uint8(3)) || (escrow[_id].state == uint8(4)), "escrow should be disputed");
-		require((_decision == uint8(0)) || (_decision == uint8(1)), "decision has to be in favor of depositer or receiver");
+		require((escrow[_id].state == 3) || (escrow[_id].state == 4), "escrow should be disputed");
+		require((_decision == 0) || (_decision == 1), "decision has to be in favor of depositer or receiver");
 
-		uint256 _feeAmount = escrow[_id].funds.mul(uint256(escrow[_id].fee));
-		_feeAmount = _feeAmount.div(uint256(10000));
+		uint256 _feeAmount = escrow[_id].funds.mul(escrow[_id].fee);
+		_feeAmount = _feeAmount.div(10000);
 		Token(tokenAddress).transfer(escrow[_id].resolver, _feeAmount);
 
 		// Funds go to depositer
-		if(_decision == uint8(0)){
+		if(_decision == 0){
 			Token(tokenAddress).transfer(escrow[_id].depositer, escrow[_id].funds);
 
-			escrow[_id].state = uint8(5);
+			escrow[_id].state = 5;
 			return true;
 		}
 		// Funds go to receiver
-		if(_decision == uint8(1)){
+		if(_decision == 1){
 			Token(tokenAddress).transfer(escrow[_id].receiver, escrow[_id].funds);
 
-			escrow[_id].state = uint8(6);
+			escrow[_id].state = 6;
 			return true;
 		}
